requirements.txt
fastapi
uvicorn
neomodel
redis
instructor
pydantic
pydantic-settings
google-genai
groq
python-dotenv
